pkg/system: make XattrError linux-only
It's only produced in Linux code, so put it in a Linux-only file.
